Change `AuthorFilterSelection`’s id to be a long

+++ WordPress/src/main/java/org/wordpress/android/ui/posts/AuthorFilterSelection.kt
@@ -1,6 +1,6 @@
 package org.wordpress.android.ui.posts
 
-enum class AuthorFilterSelection(val id: Int) {
+enum class AuthorFilterSelection(val id: Long) {
     ME(id = 0), EVERYONE(id = 1);
 
     companion object {
@@ -8,7 +8,7 @@ enum class AuthorFilterSelection(val id: Int) {
         val defaultValue = EVERYONE
 
         @JvmStatic
-        fun fromId(id: Int): AuthorFilterSelection {
+        fun fromId(id: Long): AuthorFilterSelection {
             return values().firstOrNull { it.id == id } ?: defaultValue
         }
     }
